refactor(coverage): move common parts to `vitest` package
